rag_project/requirements.txt
langchain
langchain-community
langchain-core
langchain-text-splitters
langchain-openai
langchain-groq
langchain-huggingface
langchain-chroma
chromadb
streamlit
python-dotenv
pydantic
# Document Loaders & Processing
pypdf
pdf2image
pytesseract
pillow
docx2txt
python-docx
unstructured
# Embeddings
sentence-transformers
# Model providers
openai
tiktoken
google-generativeai
pymupdf4llm
langchain-google-genai
